Rework modeling of RemoveItem
RemoveItem used to be modeled in the grammar in a way leading to the interface
inheritance of "ProperyExpression extends RemoveItem".

This is semantically incorrect so modeling of RemoveItem has been changed to
two sub-interfaces RemoveItemLabel and RemoveItemProperty representing
the removal of labels and properties, respectively.

By breaking the "ProperyExpression extends RemoveItem" inheritance path,
this cleans the Expression interface: `variable` and `nodeLabels` attributes
of RemoveItem used to pollute Expression along the inheritance path
"Expression extends ProperyExpression extends RemoveItem"